partner_chains_plutus_data::permissioned_candidates

Function permissioned_candidates_to_plutus_data

pub fn permissioned_candidates_to_plutus_data(
    candidates: &[PermissionedCandidateData],
) -> PlutusData

Converts a list of [PermissionedCandidateData] values to VersionedGenericDatum encoded as [PlutusData]. Version 0 is used for specific set of Partner Chains Key: partner chains key, AURA, Grandpa If other set of key is used, then version 1 is used. Encoding:

  VersionedGenericDatum:
  - datum: Constr 0 []
  - appendix:
    [
      [ candidates[0].sidechain_public_key
      , candidates[0].aura_public_key
      , candidates[0].grandpa_public_key
      ]
    ,
      [ candidates[1].sidechain_public_key
      , candidates[1].aura_public_key
      , candidates[1].grandpa_public_key
      ]
      // etc.
    ]
  - version: 0
or:
  VersionedGenericDatum:
  - datum: Constr 0 []
  - appendix:
    [
      [ candidates[0].sidechain_public_key
	  ,
        [
	      [ candidates[0].keys[0].id,
          , candidates[0].keys[0].bytes
          ]
        , [ candidates[0].keys[1].id,
          , candidates[0].keys[1].bytes
          ]
          // etc.
        ]
      ]
    ,
      [ candidates[1].sidechain_public_key
	  ,
        [
	      [ candidates[1].keys[0].id,
          , candidates[1].keys[0].bytes
          ]
        , [ candidates[1].keys[1].id,
          , candidates[1].keys[1].bytes
          ]
          // etc.
        ]
      ]
      // etc.
    ]
  - version: 1